requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
-e git+https://github.com/openvinotoolkit/anomalib.git@42b3ad5287f8e0833e21fbf43e0972f2439aa1ca#egg=anomalib
antlr4-python3-runtime==4.9.3
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.1.1

cycler==0.12.1
docstring_parser==0.16
einops==0.8.0
filelock==3.16.1
fonttools==4.54.1
FrEIA==0.2
frozenlist==1.5.0
fsspec==2024.10.0
ftfy==6.3.1
huggingface-hub==0.26.1
idna==3.10
imageio==2.36.0
imgaug==0.4.0
importlib_resources==6.4.5
Jinja2==3.1.4
joblib==1.4.2
jsonargparse==4.33.2
kiwisolver==1.4.7
kornia==0.7.3
kornia_rs==0.1.5
lazy_loader==0.4
lightning==2.4.0
lightning-utilities==0.11.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
open_clip_torch==2.28.0
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.3
pillow==11.0.0
propcache==0.2.0
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytorch-lightning==2.4.0
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rich==13.9.3
rich-argparse==1.5.2
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
shapely==2.0.6
six==1.16.0
sympy==1.13.1
threadpoolctl==3.5.0
tifffile==2024.9.20
timm==1.0.11
torch==2.5.0
torchmetrics==1.5.1
torchvision==0.20.0
tqdm==4.66.5
triton==3.1.0
typeshed_client==2.7.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
yarl==1.16.0
